# Remove as much of Python from JIT hot path as possible