This patch reduces the amount of Python code we have to hit when actually calling JIT-compiled modules or functions to 0, bringing some nice speedups.

Time for single application of a JIT-fused LSTM cell in a loop for 512 steps (forward only - with autograd enabled):
- before - 35us
- after - 26us
- cuDNN (nn.LSTM) - 28us?? The possible reasons for this might be that the JIT cell was specialized to never run backward, so it can save a bunch of memory bandwidth, or I did sth wrong in my benchmarking script). However, cuDNN is the only one of these which is not CPU-bound (only 15us on CPU).